requirements.txt
# FastProxy - FastAPI Reverse Proxy Requirements

# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
pyyaml==6.0.2

# Optional dependencies for production
python-multipart==0.0.12

# Testing dependencies
pytest==8.3.3
pytest-asyncio==0.24.0

